PR-SUG: rename 'CloudOrEventGridEvent' > 'Event'
