+++ src/Arcus.EventGrid.Tests.Unit/Security/WebhookValidationTests.cs
@@ -3,9 +3,9 @@ using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using Arcus.EventGrid.Parsers;
-using Arcus.EventGrid.Security.Contracts.Events.v1;
 using Arcus.EventGrid.Tests.InMemoryApi;
 using Arcus.EventGrid.Tests.Unit.Artifacts;
+using Microsoft.Azure.EventGrid.Models;
 using Microsoft.Owin.Testing;
 using Newtonsoft.Json.Linq;
 using Xunit;
@@ -19,7 +19,7 @@ namespace Arcus.EventGrid.Tests.Unit.Security
         public async Task Validate_HasValidEvent_ShouldSucceed()
         {
             // Arrange
-            var gridMessage = EventGridParser.Parse<SubscriptionValidation>(EventSamples.SubscriptionValidationEvent);
+            var gridMessage = EventGridParser.ParseFromData<SubscriptionValidationEventData>(EventSamples.SubscriptionValidationEvent);
 
             // Act
             using (var server = TestServer.Create<InMemoryTestApiStartup>())
@@ -38,9 +38,14 @@ namespace Arcus.EventGrid.Tests.Unit.Security
 
                 if (response.IsSuccessStatusCode)
                 {
-                    var responseObject = JObject.Parse(await response.Content.ReadAsStringAsync());
-                    Assert.NotNull(responseObject["validationResponse"]);
-                    Assert.Equal(gridMessage.Events[0]?.Data?.ValidationCode, responseObject["validationResponse"]);
+                    string json = await response.Content.ReadAsStringAsync();
+                    JObject responseObject = JObject.Parse(json);
+                    JToken validationResponse = responseObject["validationResponse"];
+                    Assert.NotNull(validationResponse);
+                    
+                    object data = gridMessage.Events[0]?.Data;
+                    JObject eventData = JObject.FromObject(data);
+                    Assert.Equal(eventData["validationCode"], validationResponse);
                 }
             }
         }
